major refactor - tile slider uses element.scroll methods instead of href identifiers to scroll
