+++ client/src/App.jsx
@@ -5,7 +5,6 @@ import faker from 'faker';
 import s from './styles/App.css';
 
 // import child component(s)
-import Slider from './components/Slider';
 import TileSlider from './components/TileSlider';
 import ProductTile from './components/ProductTile';
 import ProductInfo from './components/ProductInfo';
@@ -50,18 +49,18 @@ class App extends React.Component {
   render() {
     return (
       <div className="app-render">
-        <Slider
-          className="tile-slider-container"
-          selectedProduct={this.state.selectedProduct}
-          relatedProducts={this.state.relatedProducts}
-          currentPage={this.state.currentPage}
-          setCurrentPage={this.setCurrentPage.bind(this)} />
-        {/* <TileSlider
+        {/* <Slider
           className="tile-slider-container"
           selectedProduct={this.state.selectedProduct}
           relatedProducts={this.state.relatedProducts}
           currentPage={this.state.currentPage}
           setCurrentPage={this.setCurrentPage.bind(this)} /> */}
+        <TileSlider
+          className="tile-slider-container"
+          selectedProduct={this.state.selectedProduct}
+          relatedProducts={this.state.relatedProducts}
+          currentPage={this.state.currentPage}
+          setCurrentPage={this.setCurrentPage.bind(this)} />
         <ProductInfo
           selectedProduct={this.state.selectedProduct}
           relatedProducts={this.state.relatedProducts} />
