+++ client/src/components/ProductTile.jsx
@@ -5,19 +5,19 @@ import faker from 'faker';
 import s from '../styles/ProductTile.css';
 
 // import child component(s)
-import ProductReviews from './ProductReviews';
+import ReviewStars from './ReviewStars';
 
 
-const ProductTile = ({ product, index }) => {
+const ProdTile = ({ product, index }) => {
   return (
-    <li className="slide" id={'slide-' + (index + 1)}>
+    <div className="tile-slide">
       <div className="tile-image">
         <div className="tile-flag">
           <span id="reduced-flag">Reduced Price</span>
           {/* <span id={product.special + '-flag'}>{product.special}</span> */}
         </div>
-        {/* <img src={"https://hrr50-fec-images.s3-us-west-1.amazonaws.com/images/img_" + (index + 1) + ".jpeg"} alt="" style={{maxHeight: '168.66px', maxWidth: '168.66px'}} /> */}
-        <img src="http://placeimg.com/640/480/videogames" alt="" style={{maxHeight: '168.66px', maxWidth: '168.66px'}} />
+        <img src={"https://hrr50-fec-images.s3-us-west-1.amazonaws.com/images/img_" + (index + 1) + ".jpeg"} alt="" style={{maxHeight: '168.66px', maxWidth: '168.66px'}} />
+        {/* <img src="http://placeimg.com/640/480/videogames" alt="" style={{maxHeight: '168.66px', maxWidth: '168.66px'}} /> */}
         {/* <img style={{maxHeight: '168.66px', maxWidth: '168.66px'}} src={photo} alt=""/> */}
         {/* <img style={{maxHeight: '168.66px', maxWidth: '168.66px'}} src={product.images[0]} alt=""/> */}
       </div>
@@ -27,7 +27,7 @@ const ProductTile = ({ product, index }) => {
         <div className="tile-content-title">
           <p>hahah oh yea i ffasdfsjfaskld</p>
         </div>
-        <ProductReviews
+        <ReviewStars
           reviewAvg={product.reviews_avg}
           reviewTotal={product.reviews_total}
         />
@@ -36,8 +36,8 @@ const ProductTile = ({ product, index }) => {
           <p className="writing-black">{"$" + product.price}</p>
         </div>
       </div>
-    </li>
+    </div>
   );
 };
 
-export default ProductTile;
+export default ProdTile;
