+++ client/src/styles/App.css
@@ -23,21 +23,15 @@ html {
 }
 
 body {
-  /* dark mode */
-  /* background-color: rgb(36, 32, 36);
-  color: whitesmoke; */
-
-
-  /* walmart uses this font + letter spacing */
   font-family: BogleRegular;
   letter-spacing: -.02rem;
+  line-height: 1.5rem;
 
   max-width: 1360px;
   margin: auto;
 }
 
 .app-render {
-  margin-top: 1rem;
   display: grid;
 }
 
@@ -49,5 +43,8 @@ body {
   font-family: BogleBlack;
 }
 
+.hidden-true {
+  display: none;
+}
