+++ client/src/styles/ProductTile.css
@@ -1,24 +1,28 @@
-.slide {
+.tile-slide {
   max-width: 211px;
-  height: 385px;
+  height: 384.66px;
   flex-shrink: 0;
   margin: 0rem .5rem;
 
+  position: relative;
   display: flex;
   flex-direction: column;
 }
 
 .tile-image {
-  /* min-height: 55%; */
-  min-height: 220.656px;
-  min-width: 216.656px;
-  padding: 1.75rem 1.5rem 1.5rem;
-
-  object-fit: contain;
-  position: relative;
+  max-height: 220.656px;
+  max-width: 216.656px;
+  padding: 1.75rem 1.3rem 1.3rem;
 }
 
-/* * ProductTile Styling */
+.tile-flag {
+  position: absolute;
+  top: 0px;
+  left: 0px;
+
+  line-height: 1.25rem;
+  font-size: .75rem;
+}
 
 #rollback-flag {
   padding: 0rem .5rem;
@@ -32,15 +36,6 @@
   background-color: rgb(230, 241, 252);
 }
 
-.tile-flag {
-  position: absolute;
-  top: 0px;
-  left: 0px;
-
-  display: inline-flex;
-  line-height: 1.25rem;
-  font-size: .75rem;
-}
 
 .tile-content-label {
   color: #949499;
