--- client/src/styles/Slider.css
@@ -1,81 +1,0 @@
-#tile-slider-container {
-  overflow: hidden;
-  position: relative;
-
-  font-size: .875rem;
-  line-height: 1.5rem;
-}
-
-#tile-slider-header {
-  grid-row: 1;
-  padding: 2rem 0rem;
-}
-
-#tile-slider {
-  grid-row: 2;
-
-  width: 100%;
-  display: flex;
-  overflow-x: hidden;
-  overflow-y: hidden;
-  list-style-type: none;
-  scroll-behavior: smooth;
-}
-
-.hidden-true {
-  display: none;
-}
-
-.slide-pagination-left > img {
-  cursor: pointer;
-  position: absolute;
-  top: 58%;
-  width: 40px;
-  height: 40px;
-  border-radius: 100%;
-  margin-right: 35px;
-  background-color: black;
-  color: whitesmoke;
-  font-size: 1.8rem;
-}
-
-.slide-pagination-right > img {
-  cursor: pointer;
-  position: absolute;
-  top: 58%;
-  width: 40px;
-  height: 40px;
-  border-radius: 100%;
-  right: 0px;
-  margin-right: 35px;
-  background-color: black;
-  color: whitesmoke;
-  font-size: 1.8rem;
-}
-
-.slide-pagination-bottom {
-  grid-row: 3;
-  min-height: 58px;
-  padding-top: 1.5rem;
-  padding-bottom: 1rem;
-  border-bottom: 2px solid rgba(230, 231, 232);
-
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-.pagination-circle-true {
-  color: black;
-  margin: 0rem .2rem;
-  max-height: 7px;
-  pointer-events: none;
-}
-
-.pagination-circle-false {
-  color: rgb(155, 155, 155);
-  margin: 0rem .2rem;
-  max-height: 7px;
-  cursor: pointer;
-}
-
