+++ client/src/styles/TileSlider.css
@@ -1,33 +1,24 @@
-.tile-slider-container {
-  grid-row: 1;
+#tile-slider-container {
   overflow: hidden;
   position: relative;
 
   font-size: .875rem;
-  line-height: 1.5rem;
 }
 
-.image-slider-header {
+#tile-slider-header {
+  grid-row: 1;
   padding: 2rem 0rem;
+  font-size: 1rem;
 }
 
-.image-slider-contain {
+#tile-slider {
   grid-row: 2;
-  /* overflow: hidden; */
-}
 
-.slider {
   width: 100%;
-  height: 100%;
   display: flex;
   overflow-x: hidden;
-  overflow-y: hidden;
-  list-style-type: none;
+  /* overflow-y: hidden; */
   scroll-behavior: smooth;
-}
-
-.hidden-true {
-  display: none;
 }
 
 .slide-pagination-left > img {
